+++ src/call_context.rs
@@ -8,9 +8,6 @@ use canonical::{Canon, Source};
 use dusk_abi::{ContractState, Query, ReturnValue, Transaction};
 use tracing::{trace, trace_span};
 use wasmer::{Exports, ImportObject, Instance, LazyInit, Module, NativeFunc};
-use wasmer_middlewares::metering::{
-    get_remaining_points, set_remaining_points, MeteringPoints,
-};
 
 use crate::contract::ContractId;
 use crate::env::Env;
@@ -143,7 +140,6 @@ impl<'a> CallContext<'a> {
             }
 
             instance = Instance::new(&module, &import_object)?;
-            set_remaining_points(&instance, gas_meter.left());
 
             let mut memory = WasmerMemory {
                 inner: LazyInit::new(),
@@ -163,13 +159,8 @@ impl<'a> CallContext<'a> {
             instance.exports.get_native_function("q")?;
 
         let r = run_func.call(0);
-        match self.gas_reconciliation(&instance) {
-            Ok(gas) => *gas_meter = gas,
-            Err(e) => {
-                gas_meter.set_left(0);
-                return Err(e);
-            }
-        }
+        *gas_meter = self.gas_reconciliation()?;
+
         trace!(
             "Finished query with gas limit/spent: {}/{}",
             gas_meter.limit(),
@@ -225,7 +216,6 @@ impl<'a> CallContext<'a> {
                 );
             }
             instance = Instance::new(&module, &import_object)?;
-            set_remaining_points(&instance, gas_meter.left());
 
             let mut memory = WasmerMemory {
                 inner: LazyInit::new(),
@@ -243,13 +233,8 @@ impl<'a> CallContext<'a> {
         let run_func: NativeFunc<i32, ()> =
             instance.exports.get_native_function("t")?;
         let r = run_func.call(0);
-        match self.gas_reconciliation(&instance) {
-            Ok(gas) => *gas_meter = gas,
-            Err(e) => {
-                gas_meter.set_left(0);
-                return Err(e);
-            }
-        }
+        *gas_meter = self.gas_reconciliation()?;
+
         trace!(
             "Finished transact with gas limit/spent: {}/{}",
             gas_meter.limit(),
@@ -339,21 +324,7 @@ impl<'a> CallContext<'a> {
     }
 
     /// Reconcile the gas usage across the stack.
-    fn gas_reconciliation(
-        &mut self,
-        instance: &Instance,
-    ) -> Result<GasMeter, VMError> {
-        let remaining = get_remaining_points(instance);
-        match remaining {
-            MeteringPoints::Remaining(r) => {
-                self.gas_meter_mut().set_left(r as u64)
-            }
-            MeteringPoints::Exhausted => {
-                self.gas_meter_mut().set_left(0);
-                return Err(VMError::OutOfGas);
-            }
-        }
-
+    fn gas_reconciliation(&mut self) -> Result<GasMeter, VMError> {
         // If there is more than one [`StackFrame`] on the stack, then the
         // gas needs to be reconciled.
         if self.stack.len() > 1 {